requirements.txt
beautifulsoup4
lxml
requests
git+https://github.com/nmcassa/letterboxdpy.git
discord.py